% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blis_manifest.R
\name{BLIS.manifest.prepare}
\alias{BLIS.manifest.prepare}
\title{Convert a BLIS COVAIL manifest (.xlsx) to a \link[data.table]{data.table}}
\usage{
BLIS.manifest.prepare(
  BLIS.SpecimenDetail.filepath,
  manifest.physical.filepath = NULL
)
}
\arguments{
\item{BLIS.SpecimenDetail.filepath}{path to an exported (.xlsx) BLIS COVAIL manifest; the manifest should have individual vials/rows.}

\item{manifest.physical.filepath}{path to physical box/storage mappings for COVAIL samples.}
}
\value{
a \link[data.table]{data.table}
}
\description{
A BLIS COVAIL manifest -- individual vials/rows -- is converted to a \link[data.table]{data.table} along with some formatting, renaming, and re-encoding/alias columns. Additional physical box locations are added.
}
\examples{
manifest.BLIS.path<-list.files(
system.file("extdata",package="COVAILworkflow"),
pattern="SpecimenDetail", full.names = TRUE
)

manifest.physical.path<-list.files(
system.file("extdata",package="COVAILworkflow"),
pattern="physical", full.names = TRUE
)

BLIS.manifest.prepare(manifest.BLIS.path)[]
BLIS.manifest.prepare(manifest.BLIS.path,manifest.physical.path)[]

}
