% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COVAILworkbook_initialize.R
\name{col.widths.vector}
\alias{col.widths.vector}
\title{Get a numeric vector for setting Excel sheet column widths}
\usage{
col.widths.vector(dt, width.padding = 2)
}
\arguments{
\item{dt}{\code{data.frame}/\code{data.table}}

\item{width.padding}{Numeric; default 2. Adds padding to determined column widths.}
}
\value{
Numeric vector; values indicate optimal Excel sheet column widths for printing/readability.
}
\description{
Get a numeric vector for setting Excel sheet column widths
}
