% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COVAILworkbook_cellcounts.R
\name{COVAILworkbook.cellcounts}
\alias{COVAILworkbook.cellcounts}
\title{Add cell counts and a plate-based well map to an existing COVAIL workbook}
\usage{
COVAILworkbook.cellcounts(
  COVAILworkbook.path,
  cellometer.counts.path,
  rest.volume.ml = 2.2
)
}
\arguments{
\item{COVAILworkbook.path}{file.path (character string); pre-existing COVAILworkbook (.xlsx) as returned from \link{COVAILworkbook.initialize}. The COVAILworkbook file will be updated with two new worksheets.}

\item{cellometer.counts.path}{file.path (character string); Cellometer K2 cell counts .csv file.}

\item{rest.volume.ml}{Numeric; default 2.2. The total volume (mL) used to resuspend and rest cells; the total volume the cell count aliquot was taken from.}
}
\value{
an updated COVAILworkbook (.xlsx) that includes cell counts and associated derivatives (viability, conditions, cells per condition, etc.); also includes a plate-based well map for setting up stimulation conditions.
}
\description{
Add cell counts and a plate-based well map to an existing COVAIL workbook
}
\examples{
#a batch-assigned manifest as prepared and saved by:
#COVAILworkflow::BLIS.manifest.prepare()
#COVAILworkflow::manifest.batch.assign()

COVAIL_manifest_batched.path<-list.files(
  system.file("extdata",package="COVAILworkflow"),
  pattern="COVAIL_manifest_batched", full.names = TRUE
)

invisible(
  file.remove(
    list.files(
      tempdir(),
      full.names = TRUE,
      pattern = "COVAIL_[0-9]{3}_workflow.xlsx")
  )
)

#a COVAILworkbook; initialized; "COVAIL_001"
COVAILworkbook.initialize(
  covail.manifest.batched.xlsx = COVAIL_manifest_batched.path,
  batch.name = "COVAIL_001",
  workbook.dir = tempdir()
)

#update the workbook
COVAILworkbook.cellcounts(
  COVAILworkbook.path = list.files(
    tempdir(),
    full.names = TRUE,
    pattern = "COVAIL_001_workflow.xlsx"
  ),
  cellometer.counts.path = list.files(
    system.file("extdata",package="COVAILworkflow"),
    full.names = TRUE,
    pattern="cellometer_counts",
  ),
  rest.volume.ml = 2.2
)

list.files(tempdir(),pattern = "COVAIL_[0-9]{3}_workflow.xlsx")
# shell.exec(list.files(tempdir(),pattern = "COVAIL_[0-9]{3}_workflow.xlsx",full.names = TRUE))
}
