% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_protocol.R
\name{get.protocol}
\alias{get.protocol}
\title{Get a COVAIL protocol}
\usage{
get.protocol(protocol.name = c("staining.cytokine"))
}
\arguments{
\item{protocol.name}{Character string; an alias name for a particular COVAIL protocol:
\itemize{
\item 'staining.cytokine'
}}
}
\value{
A pre-formatted .html protocol
}
\description{
Get a COVAIL protocol and open the .html file (knit from R Markdown .Rmd) in a default browser window; the protocol can be directly printed from the browser window and will include per-page headers and footers.
}
