% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest_batch_assign.R
\name{manifest.batch.assign}
\alias{manifest.batch.assign}
\title{Add batch assignments to a BLIS COVAIL manifest}
\usage{
manifest.batch.assign(manifest, vials.per = 2)
}
\arguments{
\item{manifest}{a \link[data.table]{data.table} as returned from \link{BLIS.manifest.prepare}.}

\item{vials.per}{numeric; the number of sample-specific vials to assign to each batch.}
}
\value{
a list containing two \code{data.table}s; a long-data manifest and a collapsed manifest.
}
\description{
Add batch assignments to a BLIS COVAIL manifest
}
\examples{
manifest.BLIS.path<-list.files(
system.file("extdata",package="COVAILworkflow"),
pattern="SpecimenDetail", full.names = TRUE
)

manifest.physical.path<-list.files(
system.file("extdata",package="COVAILworkflow"),
pattern="physical", full.names = TRUE
)

manifest<-BLIS.manifest.prepare(manifest.BLIS.path,manifest.physical.path)[]
manifest.batch.assign(manifest,vials.per=2)

}
