% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest_cell_counts.R
\name{manifest.cell.counts}
\alias{manifest.cell.counts}
\title{Add cell counts to a manifest}
\usage{
manifest.cell.counts(
  manifest.collapsed,
  cellometer.counts.path,
  rest.volume.ml = 2.2
)
}
\arguments{
\item{manifest.collapsed}{as returned from \link{manifest.batch.assign}:
\itemize{
\item list element (if in environment)
\item sheet name (if saved as .xlsx).
}}

\item{cellometer.counts.path}{file.path (character string); Cellometer K2 cell counts .csv file.}

\item{rest.volume.ml}{Numeric; default 2.2. The total volume (mL) used to resuspend and rest cells; the total volume the cell count aliquot was taken from.}
}
\value{
an updated manifest that includes total live cells and \% viability.
}
\description{
Add cell counts to a manifest
}
\examples{
#a batch-assigned manifest as prepared by:
#COVAILworkflow::BLIS.manifest.prepare()
#COVAILworkflow::manifest.batch.assign()

COVAIL_manifest_batched.path<-list.files(
  system.file("extdata",package="COVAILworkflow"),
  pattern="COVAIL_manifest_batched", full.names = TRUE
)

COVAIL_manifest_batched<-data.table::setDT(
  openxlsx::read.xlsx(
    COVAIL_manifest_batched.path,sheet="manifest_collapsed"
  )
)

#cell counts file (.csv) from a Cellometer K2
COVAIL_cellometer_counts_path<-list.files(
  system.file("extdata",package="COVAILworkflow"),
  pattern="cellometer_counts", full.names = TRUE
)

manifest.cell.counts(
  manifest.collapsed = COVAIL_manifest_batched,
  cellometer.counts.path = COVAIL_cellometer_counts_path,
  rest.volume.ml = 2.2
)
}
