% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest_batch_assign.R
\name{plot_manifest.stage}
\alias{plot_manifest.stage}
\title{Plot a COVAIL staging box}
\usage{
plot_manifest.stage(manifest.collapsed)
}
\arguments{
\item{manifest.collapsed}{as returned from \link{manifest.batch.assign}}
}
\value{
a \link[ggplot2]{ggplot2} object
}
\description{
Plot a COVAIL staging box
}
