% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COVAILworkbook_initialize.R
\name{plot_manifest.stage}
\alias{plot_manifest.stage}
\title{Plot a COVAIL staging box}
\usage{
plot_manifest.stage(manifest.collapsed)
}
\arguments{
\item{manifest.collapsed}{as returned from \link{manifest.batch.assign}:
\itemize{
\item list element (if in environment)
\item sheet name (if saved as .xlsx).
}}
}
\value{
a \link[ggplot2]{ggplot2} object
}
\description{
Plot a COVAIL staging box
}
\examples{
#a batch-assigned manifest as prepared by:
#COVAILworkflow::BLIS.manifest.prepare()
#COVAILworkflow::manifest.batch.assign()

COVAIL_manifest_batched.path<-list.files(
  system.file("extdata",package="COVAILworkflow"),
  pattern="COVAIL_manifest_batched", full.names = TRUE
)

COVAIL_manifest_batched<-data.table::setDT(
  openxlsx::read.xlsx(
    COVAIL_manifest_batched.path,sheet="manifest_collapsed"
  )
)

stage.boxes<-plot_manifest.stage(COVAIL_manifest_batched)
stage.boxes[[1]]
stage.boxes[[41]]
}
